+++ cmd/chord/worktree_startup_test.go
@@ -34,6 +34,18 @@ func runStartupGit(t *testing.T, dir string, args ...string) {
 	}
 }
 
+func mustRunStartupGit(t *testing.T, dir string, args ...string) []byte {
+	t.Helper()
+	cmd := exec.Command("git", args...)
+	cmd.Dir = dir
+	cmd.Env = append(os.Environ(), gitTestEnvForStartup...)
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		t.Fatalf("git %s in %s: %v\n%s", strings.Join(args, " "), dir, err, out)
+	}
+	return out
+}
+
 func setupStartupRepo(t *testing.T) string {
 	t.Helper()
 	if _, err := exec.LookPath("git"); err != nil {